Updated locations
Updated the files so that functions and calls are hopefully in the correct place, trying to get __init__ and main in the correct states.
