+++ monyy/database.py
@@ -5,22 +5,14 @@ from sqlalchemy import event, func
 from sqlalchemy.engine import Engine
 from sqlite3 import Connection as SQLite3Connection
 from datetime import date
+from monyy import app, db
 
 
 
-@event.listens_for(Engine, "connect")
-def _set_sqlite_pragma(dbapi_connection, connection_record):
-    if isinstance(dbapi_connection, SQLite3Connection):
-        cursor = dbapi_connection.cursor()
-        cursor.execute("PRAGMA foreign_keys=ON;")
-        cursor.close()
 
 
 #from monyy import app
-from monyy import app
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///monyy.db'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-db = SQLAlchemy(app)
+#from monyy import app
 
 class User(UserMixin, db.Model):
     user_id = db.Column(db.Integer, primary_key=True)
@@ -367,24 +359,5 @@ def database_test():
     #print(sum_test.balance)
 
 
-db.drop_all()
-db.create_all()
-db.session.query(Account).delete()
-db.session.query(User).delete()
-db.session.query(Transaction).delete()
-db.session.query(Bank_account).delete()
-db.session.query(Transaction_bank_account).delete()
-db.session.query(Bond).delete()
-db.session.query(Transaction_bond).delete()
-db.session.query(Stock).delete()
-db.session.query(Stock_value).delete()
-db.session.query(Transaction_stock).delete()
-db.session.query(Debt).delete()
-db.session.query(Transaction_debt).delete()
-db.session.query(Real_estate).delete()
-db.session.query(Transaction_real_estate).delete()
-db.session.query(Transaction_tag).delete()
-db.session.query(Tag).delete()
-db.session.query(Account_tag).delete()
-db.session.commit()
+
 #database_test()
