+++ monyy/db_accessor.py
@@ -5,6 +5,7 @@ from sqlalchemy import event
 from sqlalchemy.engine import Engine
 from sqlite3 import Connection as SQLite3Connection
 from datetime import date
+from monyy import db
 from .database import *
 
 
@@ -23,13 +24,13 @@ class BankAccountAccessor():
         #Return that list
         return accounts
 
+    #Check that the user actually owns this account
     def ownershipCheck(self,temp_user, temp_account):
         #check that temp_user.id == temp_account.user_id
         #Raise an exception if not
         if not temp_user.user_id == temp_account.user_id:
             raise Exception("This account does not belong to the current user!")
         
-
     #get all transactions under one account
     def getAllTransactions(self,temp_user, temp_account):
         #Check that the account actually belongs to the user
@@ -45,6 +46,7 @@ class BankAccountAccessor():
         #Return the list
         return transactions
 
+    #Get a specified number of transactions based on the limite
     def getTransactions(self,temp_user, temp_account, temp_limit):
         #Check that the account actually belongs to the user
         try:
@@ -52,7 +54,21 @@ class BankAccountAccessor():
         except Exception as error: 
             raise Exception(error)
         #Make a query; joining account, transaction, bank account transaction, and bank account. Get the list of all
-        transactions = db.session.query(Account, Transaction, Transaction_bank_account, Bank_account).filter_by(account_id=temp_account.account_id).join(Transaction).order_by(Transaction.transaction_id.desc()).join(Transaction_bank_account).join(Bank_account).limit(temp_limit).all()
+        transactions = db.session.query(Account, Transaction, Transaction_bank_account, Bank_account).filter_by(account_id=temp_account.account_id).join(Transaction).order_by(Transaction.transaction_date).order_by(Transaction.transaction_id.desc()).join(Transaction_bank_account).join(Bank_account).limit(temp_limit).all()
+        #Raise an exception if they have none
+        if len(transactions) == 0:
+            raise Exception("This user has no bank accounts!")
+        #Return the list
+        return transactions
+
+    def getTransactionsOnDate(self,temp_user, temp_account, temp_limit, temp_date):
+        #Check that the account actually belongs to the user
+        try:
+            self.ownershipCheck(temp_user, temp_account)
+        except Exception as error: 
+            raise Exception(error)
+        #Make a query; joining account, transaction, bank account transaction, and bank account. Get the list of all
+        transactions = db.session.query(Account, Transaction, Transaction_bank_account, Bank_account).filter_by(account_id=temp_account.account_id).join(Transaction).filter(Transaction.transaction_date<=temp_date).order_by(Transaction.transaction_date).order_by(Transaction.transaction_id.desc()).join(Transaction_bank_account).join(Bank_account).limit(temp_limit).all()
         #Raise an exception if they have none
         if len(transactions) == 0:
             raise Exception("This user has no bank accounts!")
@@ -70,7 +86,6 @@ class BankAccountAccessor():
         query = db.session.query(func.sum(Transaction.transaction_value).label('balance')).filter_by(account_id=temp_account.account_id).filter(Transaction.transaction_id<=temp_transaction.transaction_id).filter(Transaction.transaction_date<=temp_date).first()
         #Return the int for the balance
         return int(query.balance)
-
 
     #Make a new transaction
     def makeTransaction(self,temp_user, temp_account, temp_type, temp_value, note, temp_date=date.today()):
@@ -97,8 +112,6 @@ class BankAccountAccessor():
         new_transaction_ba= Transaction_bank_account(transaction_id=new_transaction.transaction_id, bank_account_id=bank_id)
         db.session.add(new_transaction_ba)
         db.session.commit()
-
-
 
     #Make a new account
     def makeAccount(self,temp_user, temp_name, temp_value, temp_bank_name, temp_digits):
@@ -134,7 +147,6 @@ class BankAccountAccessor():
         except Exception as error:
             raise Exception("Could not create account! Error making transaction_ba! "+str(error))
 
-
     #Make withdrawal
     def makeWithdrawal(self,temp_user, temp_account, temp_value, temp_note, temp_date=date.today()):
         #Make sure the value is negative, since we are removing money
@@ -165,7 +177,9 @@ class BankAccountAccessor():
             source_value = (-1*temp_value)
             dest_value=temp_value
         elif temp_value < 0:
-            source_value=temp_value
+            source_valu
+
+            e=temp_value
             dest_value = (-1*temp_value)
         try:
             #Make a transfer transaction on the source account
@@ -175,23 +189,39 @@ class BankAccountAccessor():
         except Exception as error:
             raise Exception(error)
 
-        #Edit 4 digits
-        def editDigits(self,temp_bank_account_id, temp_digits):
-            #Get the account
-            temp_bank_account = Bank_account.query.filter_by(bank_account_id=temp_bank_account_id).first()
-            if temp_bank_account is None:
-                raise Exception("There is no bank account with that id!")
-            temp_bank_account.account_digits = temp_digits
-            db.session.commit()
+     #Edit 4 digits
+    
+    def editDigits(self,temp_bank_account_id, temp_digits):
+         #Get the account
+        temp_bank_account = Bank_account.query.filter_by(bank_account_id=temp_bank_account_id).first()
+        if temp_bank_account is None:
+            raise Exception("There is no bank account with that id!")
+        temp_bank_account.account_digits = temp_digits
+        db.session.commit()
 
-        #Edit bank name
-        def editBankName(self,temp_bank_account_id, temp_bank_name):
-            #Get the account
-            temp_bank_account = Bank_account.query.filter_by(bank_account_id=temp_bank_account_id).first()
-            if temp_bank_account is None:
-                raise Exception("There is no bank account with that id!")
-            temp_bank_account.bank_name = temp_bank_name
-            db.session.commit()  
+    #Edit bank name
+    def editBankName(self,temp_bank_account_id, temp_bank_name):
+        #Get the account
+        temp_bank_account = Bank_account.query.filter_by(bank_account_id=temp_bank_account_id).first()
+        if temp_bank_account is None:
+            raise Exception("There is no bank account with that id!")
+        temp_bank_account.bank_name = temp_bank_name
+        db.session.commit()  
+
+
+class BondAccessor():
+
+    #Get all accounts under one user
+    def getUserAccounts(self, temp_user):
+        pass
+
+    #Check that the user actually owns this account
+    def ownershipCheck(self,temp_user, temp_account):
+        pass
+    #get all transactions under one account
+    def getAllTransactions(self,temp_user, temp_account):
+        pass
+
 
 
 def BAATest():
@@ -230,3 +260,5 @@ def BAATest():
     t2 = baa.getTransactions(u,a2,1)
     print(t2[0].Transaction)
     print(baa.getBalance(u,a2,t2[0].Transaction))
+
+
