+++ monyy/main.py
@@ -1,5 +1,5 @@
-from monyy import app
-from .database import db, User
+from monyy import app, db
+from .database import *
 from .login import *
 from .db_accessor import *
 from flask import Flask, render_template, redirect
@@ -9,16 +9,20 @@ from flask_login import current_user, login_user, login_required, logout_user
 @app.route("/")
 def hello():
     #BAATest()
+    #if current_user.is_authenticated:
+    #    return "Hello "+current_user.get_username()+"!"
+    #else:
+    #    return "Hello mlemlem!"
     if current_user.is_authenticated:
-        return "Hello "+current_user.get_username()+"!"
+        return redirect("/index")
     else:
-        return "Hello mlemlem!"
+        return redirect("/login")
 
 
 @app.route("/login", methods=['GET', 'POST'])
 def login():
     if current_user.is_authenticated:
-        return redirect("/")
+        return redirect("/index")
     form = LoginForm()
     if form.validate_on_submit():
         try:
@@ -53,8 +57,22 @@ def logout():
     logout_user()
     return redirect("/login")
 
+
 @app.route("/index")
 @login_required
 def index():
+
+
+
+
+
+
+
+
+
+
+
+
+
     return render_template('index.html')
