--- monyy/models.py
@@ -1,5 +1,0 @@
-from monyy import login
-
-@login.user_loader
-def load_user(id):
-	return User.query.get(int(id))
